requirements.txt
Django==3.2.25
pytz==2024.1
redis==5.0.1
google-generativeai==0.3.2
gunicorn==21.2.0
whitenoise==6.6.0
